@@ -7,70 +7,59 @@
     "start:prod": "nx run angularlogos:serve:production",
     "build": "nx run angularlogos:build",
     "build:prod": "nx run angularlogos:build:production",
-    "ng": "nx",
-    "postinstall": "node ./decorate-angular-cli.js",
     "prettier": "pretty-quick --staged",
     "test": "nx run angularlogos:test",
     "pre-commit": "nx run angularlogos:lint && npm run prettier"
   },
   "private": false,
   "dependencies": {
-    "@angular/animations": "15.0.4",
-    "@angular/cdk": "15.0.4",
-    "@angular/common": "15.0.4",
-    "@angular/compiler": "15.0.4",
-    "@angular/core": "15.0.4",
-    "@angular/forms": "15.0.4",
-    "@angular/material": "15.0.4",
-    "@angular/platform-browser": "15.0.4",
-    "@angular/platform-browser-dynamic": "15.0.4",
-    "@angular/router": "15.0.4",
-    "@fortawesome/angular-fontawesome": "0.12.0",
-    "@fortawesome/fontawesome-svg-core": "6.2.1",
-    "@fortawesome/free-brands-svg-icons": "6.2.1",
-    "@fortawesome/free-regular-svg-icons": "6.2.1",
-    "@fortawesome/free-solid-svg-icons": "6.2.1",
-    "@nrwl/angular": "15.4.5",
+    "@angular/animations": "17.0.4",
+    "@angular/cdk": "17.0.1",
+    "@angular/common": "17.0.4",
+    "@angular/compiler": "17.0.4",
+    "@angular/core": "17.0.4",
+    "@angular/forms": "17.0.4",
+    "@angular/material": "17.0.1",
+    "@angular/platform-browser": "17.0.4",
+    "@angular/platform-browser-dynamic": "17.0.4",
+    "@angular/router": "17.0.4",
+    "@nx/angular": "17.1.3",
     "normalize.css": "8.0.1",
-    "rxjs": "7.5.7",
+    "rxjs": "7.8.1",
     "tslib": "2.4.1",
-    "zone.js": "0.11.8"
+    "zone.js": "0.14.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "15.2.10",
-    "@angular-eslint/eslint-plugin": "15.0.0",
-    "@angular-eslint/eslint-plugin-template": "15.0.0",
-    "@angular-eslint/template-parser": "15.0.0",
-    "@angular/cli": "15.0.5",
-    "@angular/compiler-cli": "15.0.4",
-    "@angular/language-service": "15.0.4",
-    "@nrwl/cypress": "15.4.5",
-    "@nrwl/eslint-plugin-nx": "15.4.5",
-    "@nrwl/jest": "15.4.5",
-    "@nrwl/linter": "15.4.5",
-    "@nrwl/workspace": "15.4.5",
-    "@types/jest": "28.1.1",
+    "@angular-devkit/build-angular": "17.0.3",
+    "@angular-devkit/core": "17.0.3",
+    "@angular-devkit/schematics": "17.0.3",
+    "@angular-eslint/eslint-plugin": "17.0.1",
+    "@angular-eslint/eslint-plugin-template": "17.0.1",
+    "@angular-eslint/template-parser": "17.0.1",
+    "@angular/compiler-cli": "17.0.4",
+    "@angular/language-service": "17.0.4",
+    "@nx/cypress": "17.1.3",
+    "@nx/eslint": "17.1.3",
+    "@nx/eslint-plugin": "17.1.3",
+    "@nx/jest": "17.1.3",
+    "@nx/workspace": "17.1.3",
+    "@schematics/angular": "17.0.3",
+    "@types/jest": "29.4.4",
     "@types/node": "16.11.7",
-    "@typescript-eslint/eslint-plugin": "5.48.0",
-    "@typescript-eslint/parser": "5.48.0",
-    "cypress": "11.2.0",
-    "eslint": "8.15.0",
-    "eslint-config-prettier": "8.1.0",
-    "eslint-plugin-cypress": "2.12.1",
-    "husky": "8.0.3",
-    "jest": "28.1.1",
-    "jest-environment-jsdom": "28.1.1",
-    "jest-preset-angular": "12.2.3",
-    "nx": "15.4.5",
+    "@typescript-eslint/eslint-plugin": "6.13.0",
+    "@typescript-eslint/parser": "6.13.0",
+    "cypress": "^12.2.0",
+    "eslint": "8.46.0",
+    "eslint-config-prettier": "9.0.0",
+    "eslint-plugin-cypress": "2.15.1",
+    "jest": "^29.5.0",
+    "jest-environment-jsdom": "^29.5.0",
+    "jest-preset-angular": "13.1.4",
+    "nx": "17.1.3",
     "prettier": "2.8.2",
     "pretty-quick": "3.1.3",
-    "ts-jest": "28.0.5",
+    "ts-jest": "29.1.1",
     "ts-node": "10.9.1",
-    "typescript": "4.8.4"
-  },
-  "husky": {
-    "hooks": {
-      "pre-commit": "npm run pre-commit"
-    }
+    "typescript": "5.2.2"
   }
 }
