@@ -1,9 +1,8 @@
 import { Routes } from '@angular/router';
-import { MetacheckComponent } from './metacheck/metacheck.component';
 import { LogosComponent } from './logos/logos.component';
 
-export const appRoutes: Routes = [
+export const routes: Routes = [
   { path: '', component: LogosComponent },
-  { path: 'metacheck', component: MetacheckComponent },
+  { path: 'metacheck', loadComponent: () => import('./metacheck/metacheck.component').then(m => m.MetacheckComponent) },
   { path: '**', redirectTo: '/' }
 ];
