@@ -1,22 +1,19 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { RouterTestingModule } from '@angular/router/testing';
 import { Component, Input, Output } from '@angular/core';
 import { MatCardModule } from '@angular/material/card';
-import { MatIconModule } from '@angular/material/icon';
-import { EMPTY, Observable, of } from 'rxjs';
+import { EMPTY, Observable } from 'rxjs';
 
 import { LogosListComponent } from './logos-list.component';
 import { DataService } from '../shared/data.service';
-
-class DataServiceMock {
-  getLogosFiltered = () => EMPTY;
-}
+import { provideNoopAnimations } from '@angular/platform-browser/animations';
 
 @Component({
   selector: 'app-search',
   template: `
     <div>{{ searchTerm | json }}</div>
-  `
+  `,
+  standalone: true
 })
 class MockSearchComponent {
   @Input() searchTerm!: string;
@@ -27,13 +24,19 @@ describe('LogosListComponent', () => {
   let component: LogosListComponent;
   let fixture: ComponentFixture<LogosListComponent>;
 
-  beforeEach(async(() => {
+  const dataServiceMock = {
+    getLogosFiltered: () => EMPTY
+  };
+
+  beforeEach(async () => {
     TestBed.configureTestingModule({
-      imports: [RouterTestingModule, MatCardModule, MatIconModule],
-      declarations: [LogosListComponent, MockSearchComponent],
-      providers: [{ provide: DataService, useClass: DataServiceMock }]
-    }).compileComponents();
-  }));
+      imports: [RouterTestingModule, MatCardModule, LogosListComponent, MockSearchComponent],
+      providers: [
+        { provide: DataService, useValue: dataServiceMock },
+        provideNoopAnimations()
+      ]
+    });
+  });
 
   beforeEach(() => {
     fixture = TestBed.createComponent(LogosListComponent);
