@@ -1,23 +1,20 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-import { Component } from '@angular/core';
-
+import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { LogosComponent } from './logos.component';
+import { RouterTestingModule } from '@angular/router/testing';
+import { NO_ERRORS_SCHEMA } from '@angular/core';
 
-@Component({
-  selector: 'app-logos-list',
-  template: ``
-})
-class LogosListComponentMock {}
 
 describe('LogosComponent', () => {
   let component: LogosComponent;
   let fixture: ComponentFixture<LogosComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(async () => {
     TestBed.configureTestingModule({
-      declarations: [LogosComponent, LogosListComponentMock]
-    }).compileComponents();
-  }));
+      imports: [LogosComponent, RouterTestingModule]
+    }).overrideComponent(LogosComponent, {
+      set: { imports: [], schemas: [NO_ERRORS_SCHEMA] }
+    });
+  });
 
   beforeEach(() => {
     fixture = TestBed.createComponent(LogosComponent);
